code queries: most visited poi query runs for regions and single poi return

+++ queries/src/main/java/hbase_schema/FriendsTable.java
@@ -31,14 +31,14 @@ public class FriendsTable implements QueriesTable {
 
 		Configuration hbaseConf = HBaseConfiguration.create();
 		HBaseAdmin admin = new HBaseAdmin(hbaseConf);
-		/*if (admin.tableExists(this.tableName)) {
+		if (admin.tableExists(this.tableName)) {
 			admin.disableTable(this.tableName);
 			admin.deleteTable(this.tableName);
-		}*/
+		}
 		HTableDescriptor descriptor = new HTableDescriptor(this.tableName);
 		descriptor.addFamily(new HColumnDescriptor("friends"));
 
-		//admin.createTable(descriptor);
+		admin.createTable(descriptor);
 		admin.close();
 		this.table = new HTable(hbaseConf, this.tableName);
 	}
